pyproject.toml
[project]
name = "car-key-sdr"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "numpy",
    "scipy",
    "matplotlib",
    "pyrtlsdr",
]
